re-create Detection class (FOR TRAINER UPLOADS ONLY)
