+++ learning_loop_node/data_classes/__init__.py
@@ -1,8 +1,9 @@
 from .annotations import AnnotationData, AnnotationEventType, SegmentationAnnotation, ToolOutput, UserInput
+from .detections import (BoxDetection, ClassificationDetection, Observation, Point, PointDetection,
+                         SegmentationDetection, Shape)
 from .general import (AnnotationNodeStatus, Category, CategoryType, Context, DetectionStatus, ErrorConfiguration,
                       ModelInformation, NodeState, NodeStatus)
-from .image_metadata import (BoxDetection, ClassificationDetection, ImageMetadata, Observation, Point, PointDetection,
-                             SegmentationDetection, Shape)
+from .image_metadata import ImageMetadata
 from .socket_response import SocketResponse
 from .training import (Errors, Hyperparameter, Model, PretrainedModel, TrainerState, Training, TrainingData,
                        TrainingError, TrainingOut, TrainingStateData, TrainingStatus)
